services/transcoding-worker/requirements.txt
# Storage Backends
boto3==1.35.36                  # S3-compatible (MinIO, StackIT)
google-cloud-storage==2.18.0    # GCS via Workload Identity